requirements.txt
Django
djangorestframework
requests
celery